chore: add saved and our team button in the header
