+++ components/Header/Header.tsx
@@ -1,6 +1,7 @@
 import { FC } from 'react'
 import Link from 'next/link'
 import clsx from 'clsx'
+import { useRouter } from 'next/router'
 
 import { ThemeToggler } from 'components/ThemeToggler/themeToggler'
 import Button from 'components/Button'
@@ -11,8 +12,14 @@ import GithubIcon from 'assets/icons/svg/github.svg'
 import LinkedinIcon from 'assets/icons/svg/linkedin.svg'
 import XIcon from 'assets/icons/svg/x.svg'
 import HeartIcon from 'assets/icons/svg/heart.svg'
+import SaveInActiveIcon from 'assets/icons/svg/nav/save-inactive.svg'
+import SaveActiveIcon from 'assets/icons/svg/nav/save-active.svg'
+import TeamInActiveIcon from 'assets/icons/svg/nav/team-inactive.svg'
+import TeamActiveIcon from 'assets/icons/svg/nav/team-active.svg'
 
 export const Header: FC = () => {
+  const router = useRouter()
+
   const iconClass =
     'fill-zinc-600 group-hover:fill-primary dark:fill-zinc-400 dark:group-hover:fill-slate-300'
   const socialIcons = [
@@ -43,15 +50,66 @@ export const Header: FC = () => {
     },
   ]
 
+  
+  const inActiveIconCls = 'stroke-gray-400'
+  const activeIconCls = 'fill-primary dark:fill-white'
+
+  const navLinks = [
+    {
+      inActiveIcon: <SaveInActiveIcon className={inActiveIconCls} />,
+      activeIcon: <SaveActiveIcon className={activeIconCls} />,
+      label: 'Saved',
+      href: '#',
+      isDisabled: true,
+    },
+    {
+      inActiveIcon: <TeamInActiveIcon className={inActiveIconCls} />,
+      activeIcon: <TeamActiveIcon className={activeIconCls} />,
+      label: 'Our Team',
+      href: '/contributors',
+      isDisabled: false,
+    },
+  ]
+
+  const renderLinks = () =>
+    navLinks.map(({ inActiveIcon, activeIcon, label, href, isDisabled }, i) => {
+      const checkRoute = (val: string) => router.asPath.startsWith(val)
+
+      const isHomeActive = !checkRoute('/saved') && !checkRoute('/contributors')
+      const isUrlMatched = checkRoute(href)
+      const isActive = label === 'Home' ? isHomeActive : isUrlMatched
+
+      return (
+        <li key={i}>
+          <a
+            href={href}
+            className={`${isDisabled ? 'cursor-not-allowed' : 'hover:bg-slate-100 hover:bg-opacity-50 dark:hover:bg-zinc-400 dark:hover:bg-opacity-10'} flex items-center justify-start p-2 gap-2 text-base font-medium leading-5 rounded-xl ${
+              isActive ? 'text-primary dark:text-white' : 'text-gray-text'
+            }`}
+          >
+            <span className="flex items-center justify-center" title={label}>
+              {isActive ? activeIcon : inActiveIcon}
+            </span>
+            <span>{label}</span>
+          </a>
+        </li>
+      )
+    })
+
   return (
     <header className="fixed top-0 left-0 z-30 row-start-1 row-end-2 h-[76px] w-screen flex items-center justify-between px-6 bg-white dark:bg-slate-800 shadow-header dark:shadow-none">
-      <div className="flex items-center justify-center">
+      <div className="flex gap-4 tall:gap-6">
         <Link href="/" aria-label="LinksHub Logo">
           <Logo />
         </Link>
       </div>
 
-      <div className="flex items-center justify-center gap-6">
+      <div className="hidden md:flex items-center justify-center gap-6">
+      <nav className="w-full flex sm:block hidden">
+          <ul className="w-full flex gap-0.5 tall:gap-1">
+            {renderLinks()}
+          </ul>
+        </nav>
         <div className="flex items-center justify-center gap-[14px]">
           {socialIcons.map(
             ({ icon, title, href, ariaLabel, showOnMobile }, i) => (
@@ -63,7 +121,7 @@ export const Header: FC = () => {
                 target="_blank"
                 rel="noopener noreferrer"
                 className={clsx(
-                  'group w-[32px] h-[32px] flex items-center justify-center rounded-md hover:bg-slate-100 dark:hover:bg-white dark:hover:bg-opacity-20 transition-colors',
+                  'group flex items-center justify-center rounded-md hover:bg-slate-100 dark:hover:bg-white dark:hover:bg-opacity-20 transition-colors',
                   showOnMobile ? '' : 'hidden sm:flex'
                 )}
               >
